requirements.txt
torch==2.2.2
transformers==4.39.3
accelerate==0.28.0
bitsandbytes==0.42.0
huggingface-hub==0.22.2
langchain==0.1.14
langchain-core==0.1.43
langchain-community==0.0.31
pypdf==4.2.0
sentence-transformers==2.6.1
beautifulsoup4==4.12.3
langserve[all]
chromadb==0.4.24
langchain-chroma==0.1.0
faiss-cpu==1.8.0
rapidocr-onnxruntime==1.3.16
unstructured==0.13.2
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv~=1.0.1
streamlit~=1.34.0
langchain-groq
pydantic~=2.7.1
tqdm~=4.66.4
langgraph~=0.0.55
typing_extensions~=4.11.0
gpt4all==2.6.0
pdfplumber~=0.11.0